requirements.txt
annotated-types==0.7.0
anyio==4.9.0
apify_client==1.9.3
apify_shared==1.4.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
distro==1.9.0
filelock==3.18.0
fsspec==2025.3.2
gdown==5.2.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.6.0
mpmath==1.3.0
networkx==3.4.2
notion-client==2.3.0
numba==0.61.2
numpy==2.2.4
openai==1.74.0
openai-whisper @ git+https://github.com/openai/whisper.git@517a43ecd132a2089d85f4ebc044728a71d49f6e
pandas==2.2.3
pydantic==2.11.3
pydantic_core==2.33.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
regex==2024.11.6
requests==2.32.3
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sympy==1.13.1
tiktoken==0.9.0
torch==2.6.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
